requirements.txt
coverage==7.10.7
attrs==22.1.0
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pluggy==1.6.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==7.1.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tomli==2.0.1
pygame~=2.3.0
XlsxWriter~=3.1.0
openpyxl~=3.1.2
